+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix JSON deserialization crash caused by unrelated warnings in the output from `xcodebuild -list -json`.
 - Retain all `@MainActor` annotated types and their constructors to workaround a bug in Swift 5.7.
 - Retain all constructors on types instantiated via `Self(...)` to workaround false positives caused by a bug in Swift.
+- `Set<AnyCancellable>` and `NSKeyValueObservation` are now included in the default values for `--retain-assign-only-property-types`.
 
 ## 2.12.0 (2023-01-13)