+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Retain all `@MainActor` annotated types and their constructors to workaround a bug in Swift 5.7.
 - Retain all constructors on types instantiated via `Self(...)` to workaround false positives caused by a bug in Swift.
 - `Set<AnyCancellable>` and `NSKeyValueObservation` are now included in the default values for `--retain-assign-only-property-types`.
+- Improve accuracy of guard-let shorthand workaround.
 
 ## 2.12.0 (2023-01-13)